feature(Line Sensor + Car Common):
Restructured Line Sensor files
Added test for Line Sensor

Added Car Dir for common functionalities
