+++ frtos/config/line_sensor_config.h
@@ -1,0 +1,16 @@
+#ifndef CONFIG_H
+#define CONFIG_H
+
+/* ADC Configuration */
+
+#define LINE_SENSOR_READ_DELAY                  ( 100 )
+
+#define LEFT_SENSOR_PIN                         ( 26 )
+#define RIGHT_SENSOR_PIN                        ( 27 )
+
+
+/* Map */
+#define MAP_START_SYMBOL                        ( 5 )
+#define MAP_SIZE 20
+
+#endif //CONFIG_H
